+++ assets/login-autofocus.js
@@ -1,0 +1,1 @@
+App.$(document).ready(function () {document.querySelector("input").focus();});